wip(redundancy_tutorials): move tutorials to subpackage
