+++ agentMET4FOF_tutorials/redundancy/redundancy_agent_four_signals.py
