+++ tests/test_tutorials.py
@@ -16,12 +16,12 @@ from agentMET4FOF_tutorials.tutorial_7_generic_metrological_agent import (
     demonstrate_metrological_stream as tut7,
 )
 
-from agentMET4FOF_tutorials.tutorial_8_redundancy_agent_four_signals import (
-    demonstrate_redundancy_agent_four_signals as tut8,
+from agentMET4FOF_tutorials.redundancy.redundancy_agent_four_signals import (
+    demonstrate_redundancy_agent_four_signals as four_signal_redundancy,
 )
 
-from agentMET4FOF_tutorials.tutorial_9_redundancy_agent_one_signal import (
-    demonstrate_redundancy_agent_onesignal as tut9,
+from agentMET4FOF_tutorials.redundancy.redundancy_agent_one_signal import (
+    demonstrate_redundancy_agent_onesignal as one_signal_redundancy,
 )
 
 
@@ -60,11 +60,11 @@ def test_tutorial_7():
     tut7().shutdown()
 
 
-def test_tutorial_8():
-    """Test executability of tutorial_8_redundancy_agent_four_signals.py."""
-    tut8().shutdown()
+def test_redundancy_agent_four_signals():
+    """Test executability of redundancy_agent_four_signals.py."""
+    four_signal_redundancy().shutdown()
 
 
-def test_tutorial_9():
-    """Test executability of tutorial_9_redundancy_agent_one_signal.py."""
-    tut9().shutdown()
+def test_redundancy_agent_one_signal():
+    """Test executability of redundancy_agent_one_signal.py."""
+    one_signal_redundancy().shutdown()
